pyproject.toml
[project]
name = "lincese-plate-detect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "onnxruntime>=1.21.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "polars>=1.26.0",
    "pycocotools>=2.0.8",
    "rfdetr>=1.0.8",
    "scikit-image>=0.25.2",
    "supervision>=0.25.1",
    "typer>=0.15.2",
]
